+++ test/unit/D_api/test_A_processor.py
@@ -18,8 +18,9 @@ import pii_extract.api.processor as mod
 from taux import auxpatch
 
 
-CONFIGFILE = Path(__file__).parents[2] / "data" / "tasklist-example.json"
-DOCUMENT = Path(__file__).parents[2] / "data" / "minidoc-example.yaml"
+DATADIR = Path(__file__).parents[2] / "data"
+CONFIGFILE = DATADIR / "tasklist-example.json"
+DOCUMENT = DATADIR /"minidoc-example.yaml"
 
 
 @pytest.fixture
@@ -302,8 +303,8 @@ def test240_tasks_detect_pii(fixture_timestamp):
 
     doc = LocalSrcDocumentFile(DOCUMENT)
     r = pd.detect(doc)
-
     assert len(r) == 2
+
     pii = list(r)
     assert str(pii[0]) == "<PiiEntity PHONE_NUMBER:+34983453999>"
     assert str(pii[1]) == "<PiiEntity CREDIT_CARD:4273 9666 4581 5642>"
@@ -321,6 +322,7 @@ def test250_tasks_detect_pii_dict(fixture_timestamp):
     r = pd.detect(doc)
 
     pii = list(r)
+    assert len(pii) == 2
     exp = {
         'detector': 1,
         'type': 'PHONE_NUMBER',
@@ -353,6 +355,73 @@ def test250_tasks_detect_pii_dict(fixture_timestamp):
         'end': 44
     }
     assert exp == pii[1].asdict()
+
+
+def test250_tasks_detect_pii_config(fixture_timestamp):
+    """
+    Test PII detection, custom config
+    """
+    config = load_config([CONFIGFILE, DATADIR / "task-config.json"])
+    pd = mod.PiiProcessor(skip_plugins=True, config=config)
+    pd.build_tasks("en")
+
+    doc = LocalSrcDocumentFile(DOCUMENT)
+    r = pd.detect(doc)
+
+    pii = list(r)
+
+    # Now, since we have deactivated context for phone numbers, there's one more
+    assert len(pii) == 3
+
+    exp = {
+        'detector': 1,
+        'type': 'PHONE_NUMBER',
+        'subtype': 'international phone number',
+        'process': {
+            'stage': 'detection'
+        },
+        'value': '+34983453999',
+        'chunkid': '3',
+        'country': 'any',
+        'lang': 'en',
+        'docid': '00000-11111',
+        'start': 44,
+        'end': 56
+    }
+    assert exp == pii[0].asdict()
+
+    exp = {
+        'detector': 2,
+        'process': {
+            'stage': 'detection'
+        },
+        'type': 'CREDIT_CARD',
+        'value': '4273 9666 4581 5642',
+        'chunkid': '4',
+        'subtype': 'standard credit card',
+        'lang': 'en',
+        'docid': '00000-11111',
+        'start': 25,
+        'end': 44
+    }
+    assert exp == pii[1].asdict()
+
+    exp = {
+        'detector': 1,
+        'type': 'PHONE_NUMBER',
+        'subtype': 'international phone number',
+        'process': {
+            'stage': 'detection'
+        },
+        'value': '+34983453000',
+        'chunkid': '5',
+        'country': 'any',
+        'lang': 'en',
+        'docid': '00000-11111',
+        'start': 37,
+        'end': 49
+    }
+    assert exp == pii[2].asdict()
 
 
 def test300_tasks_detect_chunk(fixture_timestamp):