+++ CHANGES.md
@@ -1,6 +1,6 @@
 # ChangeLog
 
-## 0.x
+## 0.7.0
  * when building tasks, pass along the config to the task constructor
  * can deactivate context filter via task config
  * use add_process_stage() from pii-data 0.5.0